[997] Prevent the cursor mouse listener to trigger unwanted update model
When the model is updated after a new version of the diagram has been
sent by the backend, the next time the cursor mouse listener is
triggered (at the next mouse move), it updates the model and thus,
forces sprotty to render the diagram again. Sometime it may look like
the diagram is blinking especially when an edge is selected and one of
its routing points has been modified.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/997
