+++ frontend/src/diagram/sprotty/DiagramServer.tsx
@@ -12,6 +12,7 @@
  *******************************************************************************/
 import {
   Bounds,
+  CursorValue,
   GQLDeletionPolicy,
   Menu,
   Palette,
@@ -140,6 +141,7 @@ export class DiagramServer extends ModelSource {
   httpOrigin;
 
   firstOpen: boolean = true;
+  getCursorOn: (element: any, diagramServer: DiagramServer) => CursorValue;
 
   initialize(registry: ActionHandlerRegistry) {
     super.initialize(registry);
@@ -368,6 +370,7 @@ export class DiagramServer extends ModelSource {
       const convertedDiagram = convertDiagram(diagram, this.httpOrigin, readOnly);
       const sprottyModel = this.modelFactory.createRoot(convertedDiagram);
       this.actionDispatcher.request<SelectionResult>(GetSelectionAction.create()).then((selectionResult) => {
+        (sprottyModel as any).cursor = 'pointer';
         const actionsToDispatch: Action[] = [UpdateModelAction.create(sprottyModel as any)];
 
         if (selectionResult.selectedElementsIDs.length > 0) {
@@ -672,5 +675,9 @@ export class DiagramServer extends ModelSource {
 
   setUpdateRoutingPointsListener(updateRoutingPointsListener: (routingPoints: Point[], edgeId: string) => void) {
     this.updateRoutingPointsListener = updateRoutingPointsListener;
+  }
+
+  setGetCursorOnListener(getCursorOn: (element, diagramServer: DiagramServer) => CursorValue) {
+    this.getCursorOn = getCursorOn;
   }
 }
