+++ frontend/src/diagram/sprotty/common/siriusCommonModule.ts
@@ -11,9 +11,11 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 import { ContainerModule } from 'inversify';
-import { configureCommand } from 'sprotty';
+import { configureCommand, TYPES } from 'sprotty';
+import { CursorMouseListener } from './CursorMouseListener';
 import { IsSiriusModelElementCommand } from './isSiriusModelElementRequest';
 
 export const siriusCommonModule = new ContainerModule((bind, unbind, isBound, rebind) => {
   configureCommand({ bind, isBound }, IsSiriusModelElementCommand);
+  bind(TYPES.MouseListener).to(CursorMouseListener).inSingletonScope();
 });
